list crud authentication
This is a basic authentication requirement for List model.

+++ Basic REST API/common/views.py
@@ -1,6 +1,7 @@
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import status
+from rest_framework import permissions
 from common.models import List
 from common.serializers import ListSerializer
 
@@ -8,6 +9,7 @@ from common.serializers import ListSerializer
 class ListAPI(APIView):
     model = List
     serializer_class = ListSerializer
+    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
 
     def get(self, request, *args, **kwargs):
         data = self.serializer_class(self.model.objects.all(), many=True)
@@ -16,7 +18,7 @@ class ListAPI(APIView):
     def post(self, request, *args, **kwargs):
         serializer = self.serializer_class(data=self.request.data)
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(owner=self.request.user)
             return Response(data=serializer.data, status=status.HTTP_201_CREATED)
         return Response(status=status.HTTP_403_FORBIDDEN)
 
@@ -24,6 +26,7 @@ class ListAPI(APIView):
 class SingleListAPI(APIView):
     model = List
     serializer_class = ListSerializer
+    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -38,6 +41,8 @@ class SingleListAPI(APIView):
         return Response(data=self.serializer_class(self.object).data, status=status.HTTP_200_OK)
 
     def put(self, request, *args, **kwargs):
+        if not self.request.user == self.object.owner:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         serializer = self.serializer_class(instance=self.object, data=self.request.data)
         if serializer.is_valid():
             serializer.save()
@@ -45,6 +50,8 @@ class SingleListAPI(APIView):
         return Response(status=status.HTTP_403_FORBIDDEN)
 
     def delete(self, request, *args, **kwargs):
+        if not self.request.user == self.object.owner:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
         self.object.delete()
         return Response(data=self.serializer_class(self.object).data, status=status.HTTP_200_OK)
